+++ forms/__init__.py
@@ -1,5 +1,5 @@
-from .auth import *
-from .event import *
-from .family import *
-from .member import *
-from .user import *
+from .auth_forms import RegisterForm, LoginForm, ResetPasswordRequestForm, ResetPasswordForm
+from .event_forms import AddEventForm
+from .family_forms import CreateFamilyForm
+from .member_forms import MemberForm
+from .user_forms import EditProfileForm
